vtc_syslog: Move VUDP API to the lib
Move all the VUDP_*() function implementation to lib/vudp.[ch] and adapt the code.
Only vtc_syslog uses this API.
